refactor : implement transaction logic
